communication/hivemind/hivemind_backend.py: fix deadlock in all_gather_object()
```
Traceback (most recent call last):
  File "/home/anon/rl-swarm/genrl-swarm/src/genrl_swarm/game/game_manager.py", line 143, in run_game
    self.run_game_round() # Loops through stages until end of round signal is received
    ^^^^^^^^^^^^^^^^^^^^^
  File "/home/anon/rl-swarm/genrl-swarm/src/genrl_swarm/game/game_manager.py", line 119, in run_game_round
    swarm_payloads = self.communication.all_gather_object(self.state.get_latest_communication()[self.rank])
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/anon/rl-swarm/genrl-swarm/src/genrl_swarm/communication/hivemind/hivemind_backend.py", line 97, in all_gather_object
    self.dht.store(
  File "/home/anon/rl-swarm/.venv/lib/python3.12/site-packages/hivemind/dht/dht.py", line 220, in store
    return future if return_future else future.result()
                                        ^^^^^^^^^^^^^^^
  File "/home/anon/rl-swarm/.venv/lib/python3.12/site-packages/hivemind/utils/mpfuture.py", line 254, in result
    return super().result(timeout)
           ^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/concurrent/futures/_base.py", line 451, in result
    self._condition.wait(timeout)
  File "/usr/lib/python3.12/threading.py", line 355, in wait
    waiter.acquire()
```

If the worker/thread that was supposed to return a result back to the DHT is dead, stuck, or resource-starved then
`result()` is just waiting forever. So in this case - `all_gather_object()` calls `dht.store()` but without any timeout,
so if in case a system is out of threads then hivemind's `store()` will be completely blocked at
`return future if return_future else future.result()`. This means it's a deadlock.

This commit adds `return_future=True` to `dht.store()` call and then wait for the future to complete within timeout defined
by `self.timeout` which is `600` secs.

Signed-off-by: 0xnightwind <nightwind1899@gmail.com>